+++ src/lib/api/get-product.ts
@@ -17,3 +17,24 @@ export async function getProductById(productId: string) {
     throw new Error(`Product with ID ${productId} not found`);
   }
 }
+
+export async function getProductBySlug(productSlug: string) {
+  const apiRoot = getApiRootSmart();
+
+  try {
+    const productResponse = await apiRoot
+      .productProjections()
+      .get({
+        queryArgs: {
+          where: `slug(en-US="${productSlug}")`,
+        },
+      })
+      .execute();
+
+    const product = productResponse.body.results[0];
+
+    return product;
+  } catch {
+    throw new Error(`Product with SLUG ${productSlug} not found`);
+  }
+}
